backend/requirements.txt
flask
flask-cors
python-dotenv
email-validator
